myAsr/requirements.txt
absl-py==0.12.0
astunparse==1.6.3
baidu-aip==2.2.18.0
beautifulsoup4==4.9.3
Brotli==1.0.9
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cycler==0.10.0
dash==1.20.0
dash-core-components==1.16.0
dash-html-components==1.1.3
dash-renderer==1.9.1
dash-table==4.11.3
docopt==0.6.2
Flask==1.1.2
Flask-Compress==1.9.0
Flask-HTTPAuth==4.3.0
flatbuffers==1.12
future==0.18.2
gast==0.4.0
gevent==21.1.2
google-auth==1.30.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
greenlet==1.0.0
grpcio==1.34.1
h5py==3.1.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
Js2Py==0.71
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.4.1
myers==1.0.1
nltk==3.5
numpy==1.19.5
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.4
Pillow==8.2.0
pipwin==0.5.1
plotly==4.14.3
pocketsphinx==0.1.15
protobuf==3.15.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio @ file:///C:/Users/%E8%8A%B1%E6%A4%92%E8%8C%B4%E9%A6%99/pipwin/PyAudio-0.2.11-cp39-cp39-win_amd64.whl
pycparser==2.20
pyjsparser==2.7.1
pyparsing==2.4.7
pypiwin32==223
PyPrind==2.11.3
PyQt5==5.15.2
pyqt5-plugins==5.15.2.2.1.0
PyQt5-Qt5==5.15.2
PyQt5-sip==12.8.1
pyqt5-tools==5.15.2.3.0.2
pySmartDL==1.3.4
python-dateutil==2.8.1
python-dotenv==0.17.0
pytz==2021.1
pywin32==225
qt5-applications==5.15.2.2.1
qt5-tools==5.15.2.1.0.1
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.6.3
simp==2.0.0
six==1.15.0
sklearn==0.0
soupsieve==2.2.1
SpeechRecognition==3.8.1
tencent-cloud-sdk==0.2.2
tencent-cloud-sdk-auth==0.2.1
tencent-cloud-sdk-common==0.1.1
tencent-cloud-sdk-core==0.2.1
tencent-cloud-sdk-serverless-database==0.1.1
tencent-cloud-sdk-serverless-functions==0.2.1
tencentcloud-sdk-python==3.0.387
tensorboard==2.5.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0rc2
tensorflow-estimator==2.5.0rc0
termcolor==1.1.0
threadpoolctl==2.1.0
tqdm==4.60.0
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.4
websocket==0.2.1
websocket-client==0.59.0
Werkzeug==1.0.1
wincertstore==0.2
wrapt==1.12.1
zope.event==4.5.0
zope.interface==5.4.0
